Switch to bdd
